requirements.txt
Django==4.2.2
requests==2.31.0
djangorestframework==3.14.0
django-filter==23.2
markdown==3.4.3
channels
daphne
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
django-cors-headers==4.1.0
django-extensions
djangochannelsrestframework
channels_redis